# backend/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    email: support@arizonagamesstore.com
    name: Поддержка
  description: |-
    API для игрового маркетплейса Arizona RP. Здесь можно купить/продать/арендовать дома, бизнесы, транспорт и всякую другую всячину. Работает на 33 серверах, поддерживает несколько валют и умеет в рейтинги продавцов.

    Основные фишки:
    - JWT авторизация (access + refresh токены)
    - Rate limiting чтобы боты не спамили
    - Email верификация
    - Загрузка картинок в AWS S3
    - Система отзывов и рейтингов
    - Жалобы на объявления
    - Автоудаление старых объявлений через 48 часов

    Сделано с душой и большим количеством кофе ☕
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Arizona Games Store API
  version: "1.0"
paths:
...
  /feedback:
    post:
      consumes:
      - application/json
      description: Оставляет отзыв о продавце. Рейтинг от 1 до 5 звезд. Отзыв нужно
        подтвердить продавцу, чтобы он отобразился
      parameters:
      - description: Данные отзыва
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Отзыв отправлен! Ждем подтверждения продавца
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный рейтинг или комментарий
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нужна авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Ты уже оставлял отзыв этому продавцу
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сохранения
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Оставить отзыв
      tags:
      - Отзывы
  /feedback/{id}/confirm:
    put:
      description: Подтверждает отзыв. Только продавец может подтвердить отзыв о себе.
        После подтверждения рейтинг пересчитывается
      parameters:
      - description: ID отзыва
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Отзыв подтвержден! Твой рейтинг обновлен
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Ты не можешь подтвердить этот отзыв
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Отзыв не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка подтверждения
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Подтвердить отзыв
      tags:
      - Отзывы
  /feedback/{nickname}:
    get:
      description: Возвращает все отзывы о продавце
      parameters:
      - description: Никнейм продавца
        in: path
        name: nickname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список отзывов
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка загрузки
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Отзывы продавца
      tags:
      - Отзывы
...
  /viewed-ads:
    get:
      description: Возвращает историю просмотренных объявлений пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Список просмотренных объявлений
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка загрузки
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: История просмотров
      tags:
      - Просмотренное
    post:
      consumes:
      - application/json
      description: Добавляет объявление в историю просмотров пользователя
      parameters:
      - description: ID объявления
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Добавлено в историю
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Не указан ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нужна авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сохранения
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавить в историю
      tags:
      - Просмотренное
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Вставь сюда JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"